Day2/pyproject.toml
[project]
name = "llm-day2"
version = "1.0.0"
description = "SK-RAG-AI Agent 교육 - Day 2"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-community",
    "langgraph>=1.0.0",
    "langgraph-cli[inmem]>=0.4.0",
    "python-dotenv>=1.0.0",
    "jupyter>=1.1.0",
    "ipykernel>=6.29.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.12.0",
    "langchain-chroma>=1.1.0",
    "langchain-tavily>=0.2.17",
    "PyMuPDF>=1.27.0"
]
